restructure code

+++ src/raycasting_utils.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   raycasting_utils.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: aweissha <aweissha@student.42.fr>          +#+  +:+       +#+        */
+/*   By: sparth <sparth@student.42heilbronn.de>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/09 14:54:12 by aweissha          #+#    #+#             */
-/*   Updated: 2024/06/09 16:32:53 by aweissha         ###   ########.fr       */
+/*   Updated: 2024/06/09 23:58:09 by sparth           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,9 +20,9 @@ double	get_lower(double a, double b)
 		return (b);
 }
 
-int	is_integer(double x)
+int	is_integer(float x)
 {
-	return (fabs(x - round(x)) < 1e-15);
+	return (fabsf(x - roundf(x)) < 1e-15);
 }
 
 int	ft_pixel(int r, int g, int b, int a)
@@ -30,7 +30,7 @@ int	ft_pixel(int r, int g, int b, int a)
 	return (r << 24 | g << 16 | b << 8 | a);
 }
 
-double	vector_len(t_vector vector)
+float	vector_len(t_vector vector)
 {
-	return (sqrt(pow(vector.x, 2) + pow(vector.y, 2)));
+	return (sqrtf(powf(vector.x, 2) + powf(vector.y, 2)));
 }
